chore: move service from /core to /service_layer
