+++ tests/unit/core/test_service.py
@@ -4,10 +4,10 @@ import pytest
 
 from src.config.config import ConfigModel
 from src.core.response_types import Task
-from src.core.service import Service
 from src.core.types import UUID, Model, TaskStatus
 from src.domain import commands
 from src.service_layer.publishing_uow import PublishingUoW
+from src.service_layer.service import Service
 from tests.integration.service_layer.fakes import FakeUoW
 from tests.unit.core.fake_handlers import FakeHandlers
 from tests.unit.core.fake_http_client import FakeHTTPClient
